fix: remove glossary from datafiller
